refactor: move getCashAccountsWithBalances to entitties
