+++ src/entities/cash-account/index.ts
@@ -4,3 +4,4 @@ export { getCashAccounts, getUserCashAccounts } from './api/get-cash-accounts';
 export { addCashAccount } from './api/add-cash-account';
 export { getCashAccount, getUserCashAccount } from './api/get-cash-account';
 export { updateCashAccount } from './api/update-cash-account';
+export { getCashAccountsWithBalances } from './api/get-cash-accounts-with-balances';
