--- src/features/cash-accounts-table/model/cash-account.mode.ts
@@ -1,9 +1,0 @@
-import { Currency } from '@prisma/client';
-
-export interface ICashAccountFromDb {
-  id: string;
-  name: string;
-  currency: Currency;
-  startBalance: number;
-  currentBalance: number;
-}
